Refactor rl algorithms core components into unified algo_core.py
