+++ tunix/rl/grpo/drgrpo_learner.py
@@ -23,19 +23,6 @@ RewardFn = rl_learner.RewardFn
 MetricFn = rl_learner.MetricFn
 
 
-@function_registry.register_advantage_estimator("drgrpo")
-def compute_advantages(rewards: jax.Array, num_generations: int) -> jax.Array:
-  """Group relative advantages -- done right.
-
-  Args:
-    rewards: reward functions output.
-    num_generations: Number of generations.
-
-  Returns:
-    Group relative advantages.
-  """
-  mean_grouped_rewards = rewards.reshape(-1, num_generations).mean(axis=1)
-  return rewards - mean_grouped_rewards.repeat(num_generations)
 
 
 @dataclasses.dataclass(slots=True, kw_only=True)
@@ -50,3 +37,7 @@ class DrGRPOConfig(grpo_learner_lib.GRPOConfig):
 
 class DrGRPOLearner(grpo_learner_lib.GrpoLearner[DrGRPOConfig]):
   """Dr. GRPO Learner."""
+
+
+from tunix.rl import algo_core
+compute_advantages = algo_core.compute_drgrpo_advantages
