+++ tunix/rl/ppo/ppo_learner.py
@@ -30,7 +30,8 @@ from tunix.rl import common
 from tunix.rl import function_registry
 from tunix.rl import rl_cluster as rl_cluster_lib
 from tunix.rl import rl_learner
-from tunix.rl.ppo import ppo_helpers
+from tunix.rl import algo_core as ppo_helpers
+
 
 TrainingInputT = rl_learner.TrainingInputT
 RewardFn = rl_learner.RewardFn
@@ -189,20 +190,19 @@ class PPOLearner(rl_learner.RLLearner[PPOConfig]):
     self.rl_cluster.actor_trainer.with_gen_model_input_fn(
         lambda x: {
             "train_example": x,
-            "epsilon_low": self.algo_config.epsilon_low,
-            "epsilon_high": self.algo_config.epsilon_high,
-            "epsilon_c": self.algo_config.epsilon_c,
-            "entropy_coef": self.algo_config.entropy_coef,
             "pad_id": self.rl_cluster.rollout.pad_id(),
             "eos_id": self.rl_cluster.rollout.eos_id(),
+            "algo_config": self.algo_config,
         }
-    )
+      )
 
     # ===== Configure the critic (value) trainer =====
+    ppo_value_loss_fn = ppo_helpers.ppo_value_loss_fn
     self.rl_cluster.critic_trainer.with_loss_fn(ppo_value_loss_fn, has_aux=True)
     self.rl_cluster.critic_trainer.with_gen_model_input_fn(
         lambda x: {
             "train_example": x,
+            "algo_config": self.algo_config,
             "clip_range_value": self.algo_config.clip_range_value,
             "pad_id": self.rl_cluster.rollout.pad_id(),
             "eos_id": self.rl_cluster.rollout.eos_id(),
@@ -219,7 +219,7 @@ class PPOLearner(rl_learner.RLLearner[PPOConfig]):
         self.algo_config.entropy_coef is not None
         and self.algo_config.entropy_coef > 0.0
     ):
-      actor_rl_metrics_to_log["loss/entropy"] = np.mean
+      actor_rl_metrics_to_log["entropy"] = np.mean
     self.rl_cluster.actor_trainer.with_rl_metrics_to_log(
         actor_rl_metrics_to_log
     )
@@ -527,154 +527,9 @@ class PPOLearner(rl_learner.RLLearner[PPOConfig]):
     super().train(train_ds, eval_ds, skip_jit)
 
 
-def ppo_value_loss_fn(
-    model: nnx.Module,
-    train_example: TrainExample,
-    clip_range_value: float | None,
-    pad_id: int,
-    eos_id: int,
-):
-  """Computes the value loss for PPO."""
-
-  prompt_ids, completion_ids, completion_mask = (
-      train_example.prompt_ids,
-      train_example.completion_ids,
-      train_example.completion_mask,
-  )
-  # ====== Loss ======
-  values = train_example.old_values
-  returns = train_example.returns
-
-  segment_ids = getattr(train_example, "segment_ids", None)
-  if segment_ids is not None:
-    # For packed sequences, prompt_ids is empty and completion_ids holds the full sequence.
-    # We predict values for token t using the model's output at t-1.
-    logits_to_keep = completion_ids.shape[1] - 1
-  else:
-    logits_to_keep = completion_ids.shape[1]
-
-  # Get new values.
-  vpreds = common.compute_score(
-      model,
-      prompt_ids,
-      completion_ids,
-      pad_id,
-      eos_id,
-      stop_gradient=False,
-      segment_ids=segment_ids,
-      segment_positions=getattr(train_example, "segment_positions", None),
-  )
-  vpreds = vpreds[:, -logits_to_keep - 1 : -1]
-
-  if segment_ids is not None:
-    # Pad the first token's value with 0.0, since it has no preceding token to predict it.
-    vpreds = jnp.pad(vpreds, ((0, 0), (1, 0)), constant_values=0.0)
-  vpred_clipped = jnp.clip(
-      vpreds, values - clip_range_value, values + clip_range_value
-  )
-  vf_losses1 = jnp.square(vpreds - returns)
-  vf_losses2 = jnp.square(vpred_clipped - returns)
-
-  clipped_vf_losses = jnp.maximum(vf_losses1, vf_losses2)
-  # "token mean" style of normalisation.
-  vf_loss = ppo_helpers.masked_mean(clipped_vf_losses, completion_mask)
-  vf_loss = 0.5 * vf_loss
-
-  aux = {
-      "vpred_mean": ppo_helpers.masked_mean(vpreds, completion_mask),
-      "vf_clipfrac": ppo_helpers.masked_mean(
-          (vf_losses2 > vf_losses1).astype(jnp.float32), completion_mask
-      ),
-  }
-  return vf_loss, aux
 
 
-@registry.register("policy_loss_fn", "ppo")
-def ppo_policy_loss_fn(
-    model: nnx.Module,
-    train_example: TrainExample,
-    epsilon_low: float,
-    epsilon_high: float,
-    epsilon_c: float | None,
-    entropy_coef: float | None,
-    pad_id: int,
-    eos_id: int,
-):
-  """Computes the policy loss for PPO."""
 
-  prompt_ids, completion_ids, completion_mask = (
-      train_example.prompt_ids,
-      train_example.completion_ids,
-      train_example.completion_mask,
-  )
-  use_dual_clip_ppo = epsilon_c is not None
-
-  # Get log probs.
-  graphdef, state = nnx.split(model)
-  per_token_logps, logits = common.compute_per_token_logps(
-      graphdef,
-      state,
-      prompt_tokens=prompt_ids,
-      completion_tokens=completion_ids,
-      pad_id=pad_id,
-      eos_id=eos_id,
-      stop_gradient=False,
-      return_logits=True,
-      segment_ids=getattr(train_example, "segment_ids", None),
-      segment_positions=getattr(train_example, "segment_positions", None),
-  )
-
-  advantages = train_example.advantages
-
-  # Compute ratio.
-  old_per_token_logps = train_example.old_per_token_logps
-  ratio = jnp.exp(per_token_logps - old_per_token_logps)
-  ratio_clipped = jnp.clip(ratio, 1 - epsilon_low, 1 + epsilon_high)
-
-  # Vanilla PPO loss
-  pg_losses_1 = -ratio * advantages
-  pg_losses_2 = -ratio_clipped * advantages
-  clip_pg_losses_1 = jnp.maximum(pg_losses_1, pg_losses_2)
-
-  # Dual-clip PPO to avoid negative-advantage policy updates
-  pg_losses = clip_pg_losses_1
-  if use_dual_clip_ppo:
-    pg_losses_3 = -epsilon_c * advantages
-    clip_pg_losses_2 = jnp.minimum(pg_losses_3, clip_pg_losses_1)
-
-    pg_losses = jnp.where(advantages < 0.0, clip_pg_losses_2, clip_pg_losses_1)
-
-    # For logging.
-    unreduced_pg_clipfrac_lower = (
-        (clip_pg_losses_1 > pg_losses_3) & (advantages < 0.0)
-    ).astype(jnp.float32)
-    pg_clipfrac_lower = ppo_helpers.masked_mean(
-        unreduced_pg_clipfrac_lower, completion_mask
-    )
-
-  # Logging
-  aux = {
-      "pg_clipfrac": ppo_helpers.masked_mean(
-          (pg_losses_2 > pg_losses_1).astype(jnp.float32), completion_mask
-      ),
-  }
-  if use_dual_clip_ppo:
-    aux["pg_clipfrac_lower"] = pg_clipfrac_lower  # pylint: disable=undefined-variable
-
-  # "token mean" style of normalisation
-  policy_loss = ppo_helpers.masked_mean(pg_losses, completion_mask)
-
-  # Compute entropy loss.
-  if entropy_coef is not None and entropy_coef > 0.0:
-    token_entropy = ppo_helpers.compute_entropy_from_logits(logits)
-    # "token mean" style of normalisation.
-    entropy_loss = ppo_helpers.masked_mean(token_entropy, completion_mask)
-    policy_loss -= entropy_coef * entropy_loss
-
-    # Logging
-    aux["loss/entropy"] = entropy_loss
-
-  return policy_loss, aux
 
 
 PpoConfig = PPOConfig
